A little more style tweaking; renamed 'Controls' to 'OrbitView'
